+++ src/headers/baseobject.h
@@ -2,14 +2,14 @@
 
 class Thing {
 public:
-	Thing(const char* texpath, SDL_Renderer* ren, double x, double y, int w, int h);
+	Thing(const char* texpath, SDL_Renderer* ren, float x, float y, int w, int h);
 	void Render(SDL_Renderer* render);
-	void ChangePos(double xnew, double ynew);
+	void ChangePos(float xnew, float ynew);
 	void ChangeDim(int wnew, int hnew);
 	void Update();
 	
-	double xpos = 0.0;
-	double ypos = 0.0;
+	float xpos = 0.0;
+	float ypos = 0.0;
 	int width = 0;
 	int height = 0;
